Setup container and route using dummy data
