+++ ordering/getters/getOrderSummaryProps.js
@@ -1,0 +1,3 @@
+import order from '../stories/mock/order'
+
+export default order
